removed wrong cwd path in wiredep task
Specifying a `cwd` option in a task's config makes that directory the "base" directory from which all the operations start.  
In this case, having the `cwd` pointing to the **app** folder, and simultaneously having**`app** in the filepaths of the subtasks, the `wiredep` was actually doubling the **app** folder, searching for files at wrong paths (like **app/app/file-name**).   
I propose this filechange in order to fix the bug and keeping this repo somewhat in sync with [the original one](https://github.com/yeoman/generator-angular/blob/master/templates/common/root/_Gruntfile.js#L177).